Fix convolution test for tf v2
Summary:
In v2, keras casts the input to the layer type, not the layer to
the input type.  So, specify the layer type too.

Manifest Task: 10908

Test Plan: All tests ok

Reviewers: #tensorflow, georgep

Reviewed By: #tensorflow, georgep

Differential Revision: https://phabricator.sourcevertex.net/D14048

+++ tensorflow/compiler/plugin/poplar/tests/conv_graph_caching_test.py
@@ -9,6 +9,7 @@ import numpy as np
 import test_utils as tu
 
 from tensorflow.compiler.tests import xla_test
+from tensorflow.python import ipu
 from tensorflow.python.platform import googletest
 from tensorflow.python.framework import constant_op
 from tensorflow.python.framework import ops
@@ -77,13 +78,15 @@ class ConvGraphCachingTest(xla_test.XLATestCase):
               2,
               1,
               use_bias=False,
-              kernel_initializer=init_ops.ones_initializer())(x)
+              kernel_initializer=init_ops.ones_initializer(),
+              dtype=np.float32)(x)
           y = math_ops.cast(y, np.float16)
           y = layers.Conv2D(
               2,
               1,
               use_bias=False,
-              kernel_initializer=init_ops.ones_initializer())(y)
+              kernel_initializer=init_ops.ones_initializer(),
+              dtype=np.float16)(y)
 
         with ops.device('cpu'):
           report = gen_ipu_ops.ipu_event_trace()
